Hack together beginnings of runner
